añadir excpeciones en msj original y reverb

+++ src/main/java/com/raio_be/raio_be/service/MensajeOriginalServiceImpl.java
@@ -6,6 +6,7 @@ import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.raio_be.raio_be.exception.MensajeOriginalNotFoundException;
 import com.raio_be.raio_be.model.MensajeOriginal;
 import com.raio_be.raio_be.repository.CategoriaRepository;
 import com.raio_be.raio_be.repository.MensajeOriginalRepository;
@@ -39,7 +40,7 @@ public class MensajeOriginalServiceImpl implements MensajeOriginalService {
     @Override
     public MensajeOriginal updateMensajeOriginal(Long id, MensajeOriginal mensajeOriginal) {
         if (!mensajeOriginalRepository.existsById(id)) {
-            return null;
+            throw new MensajeOriginalNotFoundException(id);
         }
         mensajeOriginal.setId(id);
         return mensajeOriginalRepository.save(mensajeOriginal);
@@ -47,6 +48,9 @@ public class MensajeOriginalServiceImpl implements MensajeOriginalService {
 
     @Override
     public void deleteMensajeOriginal(Long id) {
+        if (!mensajeOriginalRepository.existsById(id)) {
+            throw new MensajeOriginalNotFoundException(id);
+        }
         mensajeOriginalRepository.deleteById(id);
     }
 }
